Super Early Prototype
This is what I've been working on on the past few days, its no where near playable right now, but this provides an early look at the project and where I plan to take it. Right now, you can see the hands of two players, and it will randomly shuffle them. I hope to implement the playable functionality soon and then begin work on an AI opponent